backend/requirements.txt
Django==2.2.3
django-cors-headers==3.0.2
django-rest-framework==0.1.0
djangorestframework==3.10.2
numpy==1.17.0
pandas==0.25.0
pkg-resources==0.0.0
python-dateutil==2.8.0
pytz==2019.1
six==1.12.0
sqlparse==0.3.0
